+++ angular-ui/src/app/common/input-row/field-configuration.ts
