+++ angular-ui/src/app/common/input-row/shared/file-row.component.ts
