+++ angular-ui/src/app/common/input-row/standalone/input-row-standalone.component.html
