+++ angular-ui/src/app/common/input-row/standalone/input-row-standalone.component.ts
@@ -16,11 +16,11 @@ import {
   ReactiveFormsModule,
   Validators
 } from '@angular/forms';
-import {ErrorHandler} from "../error-handler.injectable";
+import {ErrorHandler} from "../../error-handler.injectable";
 import {Router} from "@angular/router";
 import flatpickr from 'flatpickr';
-import {InputErrorsComponent} from "./input-errors.component";
-import {FileRowComponent} from "./file-row/file-row.component";
+import {InputErrorsComponent} from "../error/input-errors.component";
+import {FileRowComponent} from "../shared/file-row.component";
 
 
 @Component({
